Form class rewritten from scratch. All forms refactored according to new Form class. Implemented form partials: added 'default' and 'text'. Implemented form conditions. Implemented form fieldsets. Implemented CkEditor element. Implemented Button element. Implemented ButtonLink element. Implemented Heading element. Implemented FileForm. Implemented EntityForm (this will allow to created forms from entity on the fly). Implemented multiple entities validation and saving support in normal form. Added Image validator (size validation). Added File MimeType validation. Added assets inline code support. Added some dochead methods for better IDE support. Added assets helper for view. Added some view functions: classof, instanceof, resolveView. Fixed some packages structure errors. Fixed exceptions double logging. Corrected some css form styling. Refactored js: added widgets, added js form support, better namespaces support (also in todo). Fixed some issues with roles and languages as json in some models. Fixed some issues founded in controller, related to forms logics. Fixed phalconeyeAuthDriver in ajaxplorer.

+++ app/modules/Core/Model/MenuItem.php
@@ -39,6 +39,9 @@ use User\Model\User;
  * @BelongsTo("parent_id", '\Core\Model\MenuItem', "id", {
  *  "alias": "MenuItem"
  * })
+ * @HasMany("id", "\Core\Model\MenuItem", "parent_id", {
+ *  "alias": "MenuItem"
+ * })
  *
  * @method static \Core\Model\MenuItem findFirst($parameters = null)
  */
@@ -263,18 +266,6 @@ class MenuItem extends AbstractModel
     }
 
     /**
-     * Prepare json string to object to interact.
-     *
-     * @return void
-     */
-    public function prepareRoles()
-    {
-        if (!is_array($this->roles)) {
-            $this->roles = json_decode($this->roles);
-        }
-    }
-
-    /**
      * Get menu item href.
      *
      * @return null|string
@@ -330,14 +321,32 @@ class MenuItem extends AbstractModel
     protected function beforeDelete()
     {
         $flag = true;
-        foreach ($this->getMenuItems() as $item) {
-            $flag = $item->delete();
-            if (!$flag) {
-                break;
+        if ($menuItems = $this->getMenuItems()) {
+            foreach ($menuItems as $item) {
+                $flag = $item->delete();
+                if (!$flag) {
+                    break;
+                }
             }
         }
 
         return $flag;
+    }
+
+    /**
+     * Spell some logic after fetching.
+     *
+     * @return void
+     */
+    protected function afterFetch()
+    {
+        if (!empty($this->roles)) {
+            $this->roles = json_decode($this->roles);
+        }
+
+        if (!empty($this->languages)) {
+            $this->languages = json_decode($this->languages);
+        }
     }
 
     /**
@@ -347,16 +356,16 @@ class MenuItem extends AbstractModel
      */
     protected function beforeSave()
     {
-        if (is_array($this->roles) && !empty($this->roles)) {
-            $this->roles = json_encode($this->roles);
-        } else {
+        if (empty($this->roles)) {
             $this->roles = null;
+        } elseif (is_array($this->roles)) {
+            $this->roles = json_encode($this->roles);
         }
 
-        if (is_array($this->languages) && !empty($this->languages)) {
-            $this->languages = json_encode($this->languages);
-        } else {
+        if (empty($this->languages)) {
             $this->languages = null;
+        } elseif (is_array($this->languages)) {
+            $this->languages = json_encode($this->languages);
         }
     }
 }
